Fix a bunch of issues that were caused by my refactor. Fix some things that were preventing th tests from running too.
