+++ common/data_refinery_common/job_lookup.py
@@ -1,4 +1,4 @@
-from enum import Enum
+from enum import Enum, unique
 from typing import List
 
 from data_refinery_common import utils
@@ -14,16 +14,105 @@ from data_refinery_common.models import (
 logger = get_and_configure_logger(__name__)
 
 
-class PipelineEnums(Enum):
-    """An abstract class to enumerate valid processor pipelines.
+@unique
+class PipelineEnum(Enum):
+    """Hardcoded pipeline names."""
 
-    Enumerations which extend this class are valid values for the
-    pipeline_required field of the Batches table.
-    """
-    pass
+    AGILENT_TWOCOLOR = "Agilent Two Color"
+    ARRAY_EXPRESS = "Array Express"
+    ILLUMINA = "Illumina"
+    NO_OP = "No Op"
+    SALMON = "Salmon"
+    SMASHER = "Smasher"
+    TXIMPORT = 'Tximport'
+    TX_INDEX = "Transcriptome Index"
+    QN_REFERENCE = "Quantile Normalization Reference"
+    JANITOR = "Janitor"
+    COMPENDIA = "Compendia"
 
 
-class ProcessorPipeline(PipelineEnums):
+@unique
+class ProcessorEnum(Enum):
+    """Hardcoded processor info in each pipeline."""
+
+    # One processor in "Agilent Two Color" pipeline
+    AGILENT_TWOCOLOR = {
+        "name": "Agilent SCAN TwoColor",
+        "docker_img": "not_available_yet",
+        "yml_file": "agilent_twocolor.yml"
+    }
+
+    # One processor in "Array Express" pipeline
+    AFFYMETRIX_SCAN = {
+        "name": "Affymetrix SCAN",
+        "docker_img": "dr_affymetrix",
+        "yml_file": "affymetrix.yml"
+    }
+
+    # One processor in "Illumina" pipeline
+    ILLUMINA_SCAN = {
+        "name": "Illumina SCAN",
+        "docker_img": "dr_illumina",
+        "yml_file": "illumina.yml"
+    }
+
+    # One processor in "No Op" pipeline
+    SUBMITTER_PROCESSED = {
+        "name": "Submitter-processed",
+        "docker_img": "dr_no_op",
+        "yml_file": "no_op.yml"
+    }
+
+    # Three processors in "Salmon" pipeline
+    SALMON_QUANT = {
+        "name": "Salmon Quant",
+        "docker_img": "dr_salmon",
+        "yml_file": "salmon_quant.yml"
+    }
+    SALMONTOOLS = {
+        "name": "Salmontools",
+        "docker_img": "dr_salmon",
+        "yml_file": "salmontools.yml"
+    }
+    TXIMPORT = {
+        "name": "Tximport",
+        "docker_img": "dr_salmon",
+        "yml_file": "tximport.yml"
+    }
+
+    # One processor in "Smasher" pipeline
+    SMASHER = {
+        "name": "Smasher",
+        "docker_img": "dr_smasher",
+        "yml_file": "smasher.yml"
+    }
+
+    # One processor in "Transcriptome Index" pipeline
+    TX_INDEX = {
+        "name": "Transcriptome Index",
+        "docker_img": "dr_transcriptome",
+        "yml_file": "transcriptome_index.yml"
+    }
+
+    QN_REFERENCE = {
+        "name": "Quantile Normalization Reference",
+        "docker_img": "dr_smasher",
+        "yml_file": "qn.yml"
+    }
+
+    COMPENDIA = {
+        "name": "Compendia Creation",
+        "docker_img": "dr_compendia",
+        "yml_file": "compendia.yml"
+    }
+
+    @classmethod
+    def has_key(cls, key):
+        """Class method that tells whether a certain key exists."""
+        return key in cls.__members__Z
+
+
+class ProcessorPipeline(Enum):
     """An enumeration of supported processors"""
     AFFY_TO_PCL = "AFFY_TO_PCL"
     AGILENT_ONECOLOR_TO_PCL = "AGILENT_ONECOLOR_TO_PCL"  # Currently unsupported
@@ -45,11 +134,6 @@ def does_processor_job_have_samples(job: ProcessorJob):
     return not (job.pipeline_applied == ProcessorPipeline.SMASHER.value \
                 or job.pipeline_applied == ProcessorPipeline.JANITOR.value \
                 or job.pipeline_applied == ProcessorPipeline.QN_REFERENCE.value)
-
-
-class DiscoveryPipeline(PipelineEnums):
-    """Pipelines which discover appropriate processing for the data."""
-    pass
 
 
 class Downloaders(Enum):
