+++ workers/data_refinery_workers/processors/utils.py
@@ -9,10 +9,9 @@ import yaml
 
 from django.conf import settings
 from django.utils import timezone
-from enum import Enum, unique
 from typing import List, Dict, Callable
 
-from data_refinery_common import job_lookup
+from data_refinery_common.job_lookup import ProcessorEnum
 from data_refinery_common.job_management import create_downloader_job
 from data_refinery_common.logging import get_and_configure_logger
 from data_refinery_common.models import (
@@ -412,104 +411,6 @@ def run_pipeline(start_value: Dict, pipeline: List[Callable]):
             return end_job(last_result, abort=True)
 
     return last_result
-
-
-@unique
-class PipelineEnum(Enum):
-    """Hardcoded pipeline names."""
-
-    AGILENT_TWOCOLOR = "Agilent Two Color"
-    ARRAY_EXPRESS = "Array Express"
-    ILLUMINA = "Illumina"
-    NO_OP = "No Op"
-    SALMON = "Salmon"
-    SMASHER = "Smasher"
-    TXIMPORT = 'Tximport'
-    TX_INDEX = "Transcriptome Index"
-    QN_REFERENCE = "Quantile Normalization Reference"
-    JANITOR = "Janitor"
-    COMPENDIA = "Compendia"
-
-
-@unique
-class ProcessorEnum(Enum):
-    """Hardcoded processor info in each pipeline."""
-
-    # One processor in "Agilent Two Color" pipeline
-    AGILENT_TWOCOLOR = {
-        "name": "Agilent SCAN TwoColor",
-        "docker_img": "not_available_yet",
-        "yml_file": "agilent_twocolor.yml"
-    }
-
-    # One processor in "Array Express" pipeline
-    AFFYMETRIX_SCAN = {
-        "name": "Affymetrix SCAN",
-        "docker_img": "dr_affymetrix",
-        "yml_file": "affymetrix.yml"
-    }
-
-    # One processor in "Illumina" pipeline
-    ILLUMINA_SCAN = {
-        "name": "Illumina SCAN",
-        "docker_img": "dr_illumina",
-        "yml_file": "illumina.yml"
-    }
-
-    # One processor in "No Op" pipeline
-    SUBMITTER_PROCESSED = {
-        "name": "Submitter-processed",
-        "docker_img": "dr_no_op",
-        "yml_file": "no_op.yml"
-    }
-
-    # Three processors in "Salmon" pipeline
-    SALMON_QUANT = {
-        "name": "Salmon Quant",
-        "docker_img": "dr_salmon",
-        "yml_file": "salmon_quant.yml"
-    }
-    SALMONTOOLS = {
-        "name": "Salmontools",
-        "docker_img": "dr_salmon",
-        "yml_file": "salmontools.yml"
-    }
-    TXIMPORT = {
-        "name": "Tximport",
-        "docker_img": "dr_salmon",
-        "yml_file": "tximport.yml"
-    }
-
-    # One processor in "Smasher" pipeline
-    SMASHER = {
-        "name": "Smasher",
-        "docker_img": "dr_smasher",
-        "yml_file": "smasher.yml"
-    }
-
-    # One processor in "Transcriptome Index" pipeline
-    TX_INDEX = {
-        "name": "Transcriptome Index",
-        "docker_img": "dr_transcriptome",
-        "yml_file": "transcriptome_index.yml"
-    }
-
-    QN_REFERENCE = {
-        "name": "Quantile Normalization Reference",
-        "docker_img": "dr_smasher",
-        "yml_file": "qn.yml"
-    }
-
-    COMPENDIA = {
-        "name": "Compendia Creation",
-        "docker_img": "dr_compendia",
-        "yml_file": "compendia.yml"
-    }
-
-    @classmethod
-    def has_key(cls, key):
-        """Class method that tells whether a certain key exists."""
-        return key in cls.__members__
 
 
 def get_os_distro():
